third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEMorphologyElement-svgdom-in-prop-expected.txt

SVG 1.1 dynamic update tests
Erode radius = 4

Tests dynamic updates of the 'in' property of the SVGFEMorphologyElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS morphologyElement.in1.baseVal is "SourceAlpha"
PASS morphologyElement.in1.baseVal is "SourceGraphic"
PASS successfullyParsed is true

TEST COMPLETE
